Add 'system' package for host code. Dual to guest/system.
