+++ guest/system/logging.go
@@ -7,7 +7,10 @@ import (
 )
 
 func init() {
-	h := slog.NewJSONHandler(os.Stderr, nil)
+	h := slog.NewJSONHandler(os.Stderr, &slog.HandlerOptions{
+		Level: slog.LevelInfo,
+	})
+
 	root := slog.New(h).With("rom", os.Args[0])
 	slog.SetDefault(root)
 }
